+++ examples/kotlin-example/src/main/kotlin/nz/bradcampbell/kotlinexample/State.kt
@@ -1,7 +1,6 @@
 package nz.bradcampbell.kotlinexample
 
 import paperparcel.PaperParcelable
-import paperparcel.Exclude
 import paperparcel.PaperParcel
 import java.util.Date
 
@@ -10,7 +9,7 @@ data class State(
     val count: Int,
     val modificationDate: Date
 ) : PaperParcelable {
-  @Exclude val somethingToExclude = 10000L
+  @Transient val somethingToExclude = 10000L
 
   companion object {
     @JvmField val CREATOR = PaperParcelState.CREATOR